@@ -1,14 +1,33 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
+import ReactDOM from 'react-dom/client';
+import { BrowserRouter } from "react-router-dom";
+import './index.scss';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { UserProvider } from './contexts/user.contexts';
+import { CategoriesProvider } from './contexts/products.context';
 
-ReactDOM.render(
+import {CartProvider } from './contexts/cart.context'
+
+import { Provider } from 'react-redux';
+import { store } from './store/store';
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
   <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+    <Provider store={store}>
+    <BrowserRouter>
+      {/* <UserProvider> */}
+        <CategoriesProvider>
+          <CartProvider>
+             <App />
+          </CartProvider>
+        </CategoriesProvider>
+      {/* </UserProvider> */}
+
+    </BrowserRouter>
+    </Provider>
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
